Fix unsatisfied references to Python docs using Intersphinx
Appendix C is essentially taken from Python documentation and contains
references to other parts of it. The paper book removes these. This change
satisfies them by linking the Python 2 documentation on-line.

We reference v2 (=v2.7) documentation as there seems not to be an
inventory file for the v2.5 documentation specifically.