Update ch 2 (DataTypes) towards the Apress version.
This change essentially replaces DataTypes.rst from Josh's repo, although
for the sake of integrity, it has been performed as a manual merge.

In order to comprehend the merge, the running text of both copies
was broken into sentences with line-breaks. As a result, even sentences
not subject to an alteration visible in the published work, are reported
as changed because they are no longer wrapped.

Bugs that prevented Josh's version building cleanly on a modern version
of Sphinx are fixed (roughly as in previous changes), and also some
minor errors in code (such as spurious word-wrapping and double-spacing,
presumably a side-effect of preparing for print.