Update ch 3 (OpsExpressPF) towards the Apress version.
This change essentially replaces OpsExpressPF.rst from Josh's repo,
as before via a manual merge to permit careful inspection. As before,
the running text of both copies was broken into one sentence per line,
to facilitate comparison.

There are substantial additions to both text and examples.

Bugs that prevented Josh's version building cleanly on a modern version
of Sphinx are fixed (roughly as in previous changes), mostly the loss of
escape characters from operators special to ReST, and a particular
problem with grid tables, the | character, and accidental substitution
references.