Update ch4 (DefiningFunctionsandUsingBuilt-Ins) towards the paper version.
We imports change from the version in Josh's repo. The *text* there is
better than we have inherited, but in terms of mark-up (particularly
marking expresions in the text as code) the incoming text appeared
to have lost them.

The usual transformation is made to one sentence per line to aid
comparison.