Fix literalincludes
Using :: to introduce a literalinclude block just puts the actual literalinclude directive in a code block instead of including the code sample. Presumably this used to work differently, but it's wrong in the hosted version over at readthedocs. 

A quick check with grep suggests that chapter 19 is the only one to use literalinclude.